+++ src/main/java/org/realityforge/replicant/client/Linkable.java
@@ -13,6 +13,11 @@ public interface Linkable
   void link();
 
   /**
+   * @return true if link() has been invoked and invalidate has not been invoked, false otherwise.
+   */
+  boolean isLinked();
+
+  /**
    * Invalidating the entity will delink the entity from related entities and remove the references to the
    * repository and change broker. After invoking this method the entity is no longer valid and should not
    * be used. This is called by the repository during delinking.